Rename userSlice to authSlice
Additional slice action renames to be more concise
